+++ atf_core/src/atf_core/__init__.py
@@ -1,7 +1,8 @@
 from atf_core.atf import ATF
-from atf_core.state_machine import StateMachine
+from atf_core.error import ATFError
 from atf_core.testblock import Testblock
 from atf_core.test import Test
 from atf_core.configuration_parser import ATFConfigurationParser
 from atf_core.recorder import ATFRecorder
 from atf_core.bagfile_helper import BagfileWriter
+from atf_core.sm_atf import SmAtfTestblock
