+++ atf_core/src/atf_core/error.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+class ATFError(Exception):
+    pass
