+++ atf_core/src/atf_core/testblock.py
@@ -3,7 +3,6 @@ import os
 import rospy
 import shutil
 
-from atf_core import StateMachine#, ATFRecorder
 from atf_msgs.msg import TestblockStatus, TestblockTrigger
 
 class Testblock:
